+++ harfbuzz/src/hb-ot.h
@@ -32,7 +32,6 @@
 
 #include "hb-ot-color.h"
 #include "hb-ot-layout.h"
-#include "hb-ot-math.h"
 #include "hb-ot-metrics.h"
 #include "hb-ot-name.h"
 #include "hb-ot-shape.h"